@@ -3,22 +3,21 @@
   "version": "0.0.0",
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.0.9",
-    "@angular/common": "^10.0.9",
-    "@angular/compiler": "^10.0.9",
-    "@angular/core": "^10.0.9",
-    "@angular/forms": "^10.0.9",
+    "@angular/animations": "10.2.0",
+    "@angular/common": "10.2.0",
+    "@angular/compiler": "10.2.0",
+    "@angular/compiler-cli": "10.2.0",
+    "@angular/core": "10.2.0",
+    "@angular/forms": "10.2.0",
     "@angular/localize": "^10.2.0",
-    "@angular/platform-browser": "^10.0.9",
-    "@angular/platform-browser-dynamic": "^10.0.9",
-    "@angular/router": "^10.0.9",
-    "@ngxs/store": "^3.7.0",
-    "bootstrap": "^4.5.2",
-    "jquery": "1.9.1 - 3",
-    "popper.js": "^1.16.1",
-    "rxjs": "^6.6.2",
-    "tslib": "^2.0.0",
-    "zone.js": "^0.10.3"
+    "@angular/platform-browser": "10.2.0",
+    "@angular/platform-browser-dynamic": "10.2.0",
+    "@angular/router": "10.2.0",
+    "@ngxs/store": "3.7.0",
+    "rxjs": "6.6.3",
+    "tslib": "2.0.3",
+    "typescript": ">=3.9 <4.1",
+    "zone.js": "0.11.2"
   },
   "scripts": {
     "ng": "ng",
